requirements.txt
fastapi
pydantic
uvicorn
python-multipart
python-dotenv
requests
aws-lambda-powertools
mangum
pynamodb
boto3
aws-xray-sdk
Pillow